merge

+++ frontends/CsoundAC/MusicModel.hpp
@@ -34,6 +34,7 @@
 #include "CppSound.hpp"
 #include "Node.hpp"
 #include "Score.hpp"
+#include <stdint.h>
 #endif
 
 namespace csound
@@ -51,7 +52,7 @@ namespace csound
     virtual ~MusicModel();
     virtual void initialize();
     virtual int generate();
-    virtual long getThis();
+    virtual intptr_t getThis();
     virtual Node *getThisNode();
     /**
      * Translate the generated score to a Csound score
